REFACTOR: worker logic moved from workers/python to /worker
