+++ apps/worker/app/domain/errors.py
