+++ apps/worker/main.py
@@ -1,7 +1,7 @@
-from app.application.process_submission import ProcessSubmission
-from app.adapters.driven.docker_executor import DockerExecutor
-from app.adapters.driven.http_api_client import HTTPAPIClient
-from app.adapters.driving.rabbitmq_consumer import RabbitMQConsumer
+from apps.worker.app.application.process_submission import ProcessSubmission
+from apps.worker.app.adapters.driven.docker_executor import DockerExecutor
+from apps.worker.app.adapters.driven.http_api_client import HTTPAPIClient
+from apps.worker.app.adapters.driving.rabbitmq_consumer import RabbitMQConsumer
 
 
 def main():
